feat: add initial support for syncing files to the enclave
The new `files` key in the manifest specifies the list of file paths that should be synced
to the enclave. Only absolute paths without any relative components are accepted.

This is a simplified implementation to support passing Kubernetes service account credentials
and other ConfigMap and Secret data mounted as pod volumes. It comes with the following
limitations:

* All files must exist on the host (pod) when `enclaver-run` starts. If a file does not exist
  when the enclave is launched, the enclave will hang forever waiting for its initial sync.
* If the path specified in a manifest is a symlink and one of path components of its target
  is removed or replaced non-atomically (i.e. removed and recreated slower than the inotify
  debounce interval of 1 second), the file will stop being synced.
* Files are synced sequentially, so the startup time of the enclave may be slow if they are large.